made eventrecorder optional; added txn status marker

+++ src/main/java/org/reactiveminds/txpipe/core/dto/Event.java
@@ -15,7 +15,7 @@ public class Event {
 	public void setTxnId(String txnId) {
 		this.txnId = txnId;
 	}
-	public long getEventId() {
+	public int getEventId() {
 		return eventId;
 	}
